+++ CHANGELOG.md
@@ -28,12 +28,15 @@ to [Semantic Versioning]. Full commit history is available in the
 #### Added
 
 - Add MuData Minification option to {class}`~scvi.model.TOTALVI` {pr}`3061`.
+- Add support for Numpy 2.0 {pr}`2842`.
+-
 
 #### Fixed
 
 #### Changed
 
 - Updated the CI workflow with internet, private and optional tests {pr}`3082`.
+- Changed loompy stored files to anndata {pr}`2842`.
 
 #### Removed